+++ packages/jest-runtime/src/index.ts
@@ -526,7 +526,6 @@ export default class Runtime {
               // @ts-expect-error Jest uses @types/node@16. Will be fixed when updated to @types/node@20.11.0
               meta.dirname = path.dirname(modulePath);
 
-              // @ts-expect-error It should not be async. Will be fixed when updated to @types/node@20.11.0
               meta.resolve = (specifier, parent = metaUrl) => {
                 const parentPath = fileURLToPath(parent);